+++ panel/version-control.rkt
@@ -1,11 +1,12 @@
 #lang racket/gui
 
-(require "../component/common-text.rkt"
+(require framework
+         "../component/common-text.rkt"
          "../project-manager.rkt")
 
 (provide version-control%)
 (define version-control%
-  (class vertical-panel%
+  (class panel:vertical-dragable%
     (super-new)
 
     ;;; commit editor
@@ -33,41 +34,52 @@
     (send editor-canvas set-editor commit-message-editor)
 
     ;;; ready/changes zone
-    (define (run cmd [callback #f])
-      (parameterize ([current-directory (current-project)])
-        (match-let ([(list out in pid err invoke) (process cmd)])
-          (invoke 'wait)
-
-          (when callback
-            (callback out in err))
-
-          (close-output-port in)
-          (close-input-port out)
-          (close-input-port err))))
-
-    (define files-zone (new group-box-panel% [parent this]
+    (define zone (new panel:vertical-dragable% [parent this]))
+    (send this set-percentages (list 1/3 2/3))
+    (define button-zone (new horizontal-panel% [parent zone]))
+    (define files-zone (new group-box-panel% [parent zone]
                             [label "files"]
                             [alignment '(left top)]))
+    (send zone set-percentages (list 1/10 9/10))
+
+    (new button% [parent button-zone]
+         [label "select all"]
+         [callback
+          (λ (btn event)
+            (for ([file-obj (send files-zone get-children)])
+              (send file-obj add-to-ready)))])
+    (new button% [parent button-zone]
+         [label "unselect all"]
+         [callback
+          (λ (btn event)
+            (for ([file-obj (send files-zone get-children)])
+              (send file-obj remove-from-ready)))])
+    (new button% [parent button-zone]
+         [label "clean up"]
+         [callback
+          (λ (btn event)
+            (run "git reset --hard")
+            (run "git clean -fd")
+            (for ([f (send files-zone get-children)])
+              (send files-zone delete-child f)))])
 
     (define/public (update-status)
       ; show current status one file one line
       (run "git status --short --untracked-files=all"
            (λ (out in err)
              (let loop ([output (read-line out)])
-               (cond
-                 [(eof-object? output) (void)]
-                 [else
-                  (match-let ([(cons kind filename) (parse-git-output output)])
-                    (new file-object% [parent files-zone]
-                         [filename filename]
-                         [add-to-ready
-                          (λ (this filename)
-                            (run (format "git add ~a" (build-path (current-project) filename))))]
-                         [remove-from-ready
-                          (λ (this filename)
-                            (run (format "git reset HEAD ~a" (build-path (current-project) filename))))]
-                         [status kind]))
-                  (loop (read-line out))])))))
+               (unless (eof-object? output)
+                 (match-let ([(cons kind filename) (parse-git-output output)])
+                   (new file-object% [parent files-zone]
+                        [filename filename]
+                        [λ-add-to-ready
+                         (λ (this filename)
+                           (run (format "git add ~a" (build-path (current-project) filename))))]
+                        [λ-remove-from-ready
+                         (λ (this filename)
+                           (run (format "git reset HEAD ~a" (build-path (current-project) filename))))]
+                        [status kind]))
+                 (loop (read-line out)))))))
 
     ;;; init
     (update-status)))
@@ -75,23 +87,42 @@
 (define file-object%
   (class horizontal-panel%
     (init-field filename
-                add-to-ready
-                remove-from-ready
+                λ-add-to-ready
+                λ-remove-from-ready
                 status)
     (super-new [alignment '(left top)])
 
-    (new check-box% [parent this]
-         [label filename]
-         [value (match status
-                  ['ready #t]
-                  ['changes #f])]
-         [callback
-          (λ (check-box event)
-            (define clicked? (send check-box get-value))
-            (displayln clicked?)
-            (if clicked?
-                (add-to-ready this filename)
-                (remove-from-ready this filename)))])))
+    (define check-box
+      (new check-box% [parent this]
+           [label filename]
+           [value (match status
+                    ['ready #t]
+                    ['changes #f])]
+           [callback
+            (λ (check-box event)
+              (define clicked? (send check-box get-value))
+              (if clicked?
+                  (λ-add-to-ready this filename)
+                  (λ-remove-from-ready this filename)))]))
+
+    (define/public (add-to-ready)
+      (λ-add-to-ready this filename)
+      (send check-box set-value #t))
+    (define/public (remove-from-ready)
+      (λ-remove-from-ready this filename)
+      (send check-box set-value #f))))
+
+(define (run cmd [callback #f])
+  (parameterize ([current-directory (current-project)])
+    (match-let ([(list out in pid err invoke) (process cmd)])
+      (invoke 'wait)
+
+      (when callback
+        (callback out in err))
+
+      (close-output-port in)
+      (close-input-port out)
+      (close-input-port err))))
 
 (define (parse-git-output output)
   (cons
@@ -106,18 +137,19 @@
    (substring output 3)))
 
 (module+ main
+  (define testing-dir (build-path (find-system-path 'home-dir) "racket.tw" "developing"))
+  (unless (directory-exists? testing-dir)
+    (error 'file "no such dir"))
+
+  (current-project testing-dir)
   (define test-frame (new frame%
                           [label "Version Control Panel"]
                           [width 300]
                           [height 600]))
 
-  (define home-dir (find-system-path 'home-dir))
-  (define testing-dir (build-path home-dir "racket.tw" "developing"))
-  (unless (directory-exists? testing-dir)
-    (error 'file "no such dir"))
+  (define vc 
+    (new version-control%
+         [parent test-frame]))
 
-  (define vc (parameterize ([current-project testing-dir])
-               (new version-control%
-                  [parent test-frame])))
-
+  (send test-frame center)
   (send test-frame show #t))