+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 ### To Release v0.4.0
 
-* [vc] c+k show commit editor
-* the following pairs would complete and wrap selected text automatically
+* [migrate] c+e run REPL
+* [vc:migrate] c+k show commit editor
+* [migrate] the following pairs would complete and wrap selected text automatically
     * `()`
     * `[]`
     * `{}`
     * `""`
-* c+; comment selected text(or that line if no selected text) if uncommented, uncomment if commented
+* [migrate] c+; comment selected text(or that line if no selected text) if uncommented, uncomment if commented
 * o+backspace delete previous sexp
 * c+backspace delete whole line from current position